requirements.txt
torch
Pillow
diffusers
accelerate
google-generativeai
tqdm